[fix] fix test on db

+++ convert_app/db/mongo_db/model.py
@@ -1,13 +1,8 @@
-from decimal import Decimal
-
-from bson import Decimal128
-
-
 class Document:
     def __init__(self, time: str, currency: str, rate: str) -> None:
         self.time = time
         self.currency = currency
-        self.rate = Decimal128(Decimal(rate))
+        self.rate = rate
 
     def to_dict(self) -> dict:
         return {"time": self.time,
